seperated grid and runner
