+++ java/com/example/demo2/MouseClick.java
@@ -1,0 +1,7 @@
+package com.example.demo2;
+
+
+public interface MouseClick {
+	public boolean reveal(Grid g);
+	public boolean flag(Grid g);
+}
